Applies changes in ts-apollo-prisma
- https://github.com/dooboolab/ts-apollo-prisma/pull/12
