+++ server/src/types/resolvers/Post/index.ts
@@ -1,0 +1,2 @@
+export * from './mutation';
+export * from './query';
