--- server/src/types/resolvers/Query.ts
@@ -1,68 +1,0 @@
-import { intArg, queryType, stringArg } from '@nexus/schema';
-
-import { getUserId } from '../../utils';
-
-export const Query = queryType({
-  definition(t) {
-    t.field('me', {
-      type: 'User',
-      nullable: true,
-      resolve: (parent, args, ctx) => {
-        const userId = getUserId(ctx);
-
-        return ctx.prisma.user.findOne({
-          where: {
-            id: userId,
-          },
-        });
-      },
-    });
-
-    t.list.field('feed', {
-      type: 'Post',
-      resolve: (parent, args, ctx) => {
-        return ctx.prisma.post.findMany({
-          where: { published: true },
-        });
-      },
-    });
-
-    t.list.field('filterPosts', {
-      type: 'Post',
-      args: {
-        searchString: stringArg({ nullable: true }),
-      },
-      resolve: (parent, { searchString }, ctx) => {
-        return ctx.prisma.post.findMany({
-          where: {
-            OR: [
-              {
-                title: {
-                  contains: searchString,
-                },
-              },
-              {
-                content: {
-                  contains: searchString,
-                },
-              },
-            ],
-          },
-        });
-      },
-    });
-
-    t.field('post', {
-      type: 'Post',
-      nullable: true,
-      args: { id: intArg() },
-      resolve: (parent, { id }, ctx) => {
-        return ctx.prisma.post.findOne({
-          where: {
-            id: Number(id),
-          },
-        });
-      },
-    });
-  },
-});
