+++ server/src/types/resolvers/User/index.ts
@@ -1,0 +1,3 @@
+export * from './mutation';
+export * from './subscription';
+export * from './query';
