+++ server/src/types/resolvers/index.ts
@@ -1,3 +1,2 @@
-export * from './Mutation';
-export * from './Query';
-export * from './Subscription';
+export * from './User';
+export * from './Post';
